UI overhaul

+++ app/Http/Controllers/CartController.php
@@ -139,7 +139,7 @@ class CartController extends Controller
     }
     //ADDS TO bought products
     public function checkout(Request $request){
-        
+        $addresses = Address::all()->where('user_id',Auth::user()->id);
         $bought = Cart::get();
         foreach($bought as $key => $value){
             BoughtProducts::create([
@@ -152,9 +152,13 @@ class CartController extends Controller
                 'user_id' => $user_id = $request->user()->id,
             ]);
         }
-        
+        if(count($addresses)>='1'){
         Cart::where('user_id', $user_id)->delete();
         return redirect('my_purchase')->with(['bought' => 'Bought successfully'], 200);
+        }
+        else{
+            return redirect()->back()->with('error', 'No address detected');
+        }
 
     }
     //BUY NOW
